Reorganized the folder structure and improved exception handling.
