+++ custom_exceptions.py
@@ -1,0 +1,6 @@
+class DownloadException(Exception):
+    pass
+
+
+class SubtitlesNotAvailableException(Exception):
+    pass
